api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
anthropic>=0.28.0
pyyaml>=6.0.1
pdfplumber>=0.11.0
python-docx>=1.1.2
openpyxl>=3.1.2
python-multipart>=0.0.9
supabase>=2.4.0
python-dotenv>=1.0.0
PyJWT[crypto]>=2.8.0
